requirements.txt
pygame==2.5.2
asyncio==3.4.3
edge-tts==1.3.5
python-dotenv==1.0.1
playsound==1.3.0
AppOpener==1.6
pywhatkit==5.4
beautifulsoup4==4.12.3
rich==13.7.1
groq==0.5.0
requests==2.31.0
keyboard==0.13.5
cohere==4.38
googlesearch-python==1.2.3
